Act010_Main.java Act010_Main_Presenter.java Act010_Main_Presenter_Impl.java     - Passagem de product_code. @#$

+++ app/src/main/java/com/namoadigital/prj001/ui/act010/Act010_Main_Presenter.java
@@ -20,7 +20,7 @@ public interface Act010_Main_Presenter {
 
     SpannableString getTagLblText(String tag_lbl, String tag_desc);
 
-    void callTicketCreationService(long customer_code, int type_code, String site_code, long operation_code, int product_code, int serial_code, String comments);
+    void callTicketCreationService(long customer_code, int type_code, String site_code, long operation_code, long product_code, int serial_code, String comments);
 
     void callTicketDownload(Integer ticketPrefix, Integer ticketCode);
